requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
pandas==2.2.0
numpy==1.26.4
psycopg2-binary==2.9.11